Add uwb code examples
Add the code with which an anchor can be setup and code with which an tag can be setup.

In one example the tag does not use the self made library and thus can only communicate with one anchor.

Also added an example using the self made library which can communicate and read the data from any anchor as used in this project. This code allows the tag to read from two anchors which results in two distances with which the location of the tag can be calculated if the distance between the two anchors is known